gfx: Remove unused settings and code
Drop S9xInitUpdate() which is only a stub here and was originally used
upstream in the Windows port until 1.52 where it was unused in the
Windows port as well. This function has already been removed in the
libretro snes9x2005 and 2010 cores.

Also drop checking the Settings.SixteenBit flag as this fork no longer
contains 8bpp graphics rendering code.